requirements.txt
opencv-contrib-python>=4.8.0.76
numpy>=1.24.0
PyYAML>=6.0.0
